xhs_semantic_service/requirements.txt
flask==3.0.3
text2vec==1.2.9
faiss-cpu==1.8.0
numpy==1.26.4
pymysql==1.1.1
